ci: build a docker image with the app at entrypoint
