+++ vhcalc/tools/chunk.py
@@ -6,6 +6,10 @@ from typing import Any, Iterator
 
 
 def chunks(iterable: Iterator[Any], size: int) -> Iterator[Any]:
+    """
+    >>> list(chunks(range(10), 3))
+    [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9]]
+    """
     iterator = iter(iterable)
     for first in iterator:
         yield list(chain([first], islice(iterator, size - 1)))
