+++ vhcalc/tools/progress_bar.py
@@ -1,3 +1,4 @@
+from datetime import timedelta
 from typing import Optional, Tuple
 
 from rich import get_console
@@ -6,12 +7,23 @@ from rich.progress import (
     BarColumn,
     DownloadColumn,
     Progress,
+    Task,
     TaskID,
     TextColumn,
+    TimeRemainingColumn,
     TransferSpeedColumn,
 )
+from rich.text import Text
 
-from vhcalc.tools.rich_colums import TimeElapsedOverRemainingColumn
+
+class TimeElapsedOverRemainingColumn(TimeRemainingColumn):
+    """Renders remaining on elapsed time."""
+
+    def render(self, task: Task) -> Text:
+        """Show times elapsed/remaining."""
+        elapsed_delta = timedelta(seconds=int(task.elapsed or -1))
+        remaining = super().render(task)
+        return Text(f"{str(elapsed_delta)}/{remaining}", style="progress.remaining")
 
 
 def configure_progress_bar(
